requirements.txt
streamlit
yfinance
pandas 
numpy
plotly
scikit-learn
ta
tensorflow
textblob
newspaper3k==0.2.8
lxml==4.9.3